moving json render to solution_api controller to avoid view_template mismatch (result of adding liveview + upgrading phx
